Add in a hard die() if an active timer is overwritten.  This is a big problem, as the old timer will still fire if it hasn't been removed previously.
Set the request's REQ_TIMER to undef in remove_timeout() so multiple
calls to it won't attempt to alarm_remove() the timer again.

Replace a $request->timer(undef) call with $request->remove_timeout().
Setting the timer to undef does not call alarm_remove(), but
remove_timeout() does this necessary magic.