[rt.cpan.org 70268] Some DNS providers redirect bad hosts.
They redirect bad hosts instead of allowing outright failure.  I think
it's supposed to be helpful, or improve domain sales, or both, or
something.  But anyway, it meant that people using certain DNS
providers wouldn't be able to cleanly test this distribution.

Thanks to "blackcode" for reporting this.  Unfortunately I'm not sure
of a foolproof solution other than to make the test accept 303s.

Patches are welcome, dear reader from the future.