Honor "Connection: close" response header by closing the connection.
Without closing the connection, POE::Component::Client::Keepalive
would put the connection back into its pool and detect closure when
the server later shut down the socket.

Nicolas Dehaine and Brad Sacks discovered this to be a problem when
they tried to make a lot of rapid connections to a single address.
The to-be-defunct connection didn't have enough time to shut down, so
it would be returned to POE::Component::Client::HTTP.  The HTTP
transaction would then fail, and an error would be reported.

Their test case allowed me to repeat the problem and diagnose it.
Otherwise this would still be broken.  Thanks, guys!